+修改 matrix 乘 matrix 由于 写成了 int 各个值都取整了，导致 的 矩阵乘法 的错误

+++ _mac_playground/ayyrender/source/math.hpp
@@ -401,7 +401,7 @@ struct Matrix
         {
             for(int col = 0;col < n;col++)
             {
-                int counter = 0;
+                float counter = 0;
                 for(int k = 0;k < n;k++)
                 {
                     counter += data[row][k] * other.data[k][col];
